Breaks Skills into its own component.
